requirements.txt
Flask==3.0.3
gevent==23.9.1
Pillow==10.1.0
Pillow==10.3.0
Werkzeug==3.0.3
